+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Link not working with `target` different from the default value.
+
 ## [8.91.3] - 2020-02-18 [YANKED]
 ### Fixed
 - Issue where page wouldn't remount if params changed within the same page.